% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outfitfct}
\alias{outfitfct}
\title{Compute fit statistics}
\usage{
outfitfct(beta, theta, data)
}
\arguments{
\item{beta}{Vector or matrix of item parameters.}

\item{theta}{Vector of person parameters.}

\item{data}{Matrix with item responses.}

\item{N}{Number of persons.}

\item{K}{Number of items.}
}
\description{
Compute fit statistics
}
