% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotfct}
\alias{plotfct}
\title{Density plot function}
\usage{
plotfct(z, probs, breaks, labels, colours, xtitle = NULL, title = NULL)
}
\arguments{
\item{z}{Vector of statistics to be plotted.}

\item{probs}{Vector of percentiles.}

\item{breaks}{Vector of breaks.}

\item{labels}{Vector of labels}

\item{colours}{Vector of colours}

\item{xtitle}{Character string for x axis title}

\item{title}{Character string for plot title}
}
\description{
Density plot function
}
