% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zeromean_confirm}
\alias{zeromean_confirm}
\title{Create a modal dialog UI when mean of item parameters is not zero}
\usage{
zeromean_confirm()
}
\arguments{
\item{...}{Elements to include}
}
\description{
Create a modal dialog UI when mean of item parameters is not zero
}
